refactor(fe): Move modal and related state/callbacks from AdminCampaignComponent to CampaignCard/Content
